docs(java): Remove low-level section about database in high level one
The section "Database Support" is high-level, describing supported
databases, etc.

However, then it mentions write-through and refers to a subsection.
That is rather confusing and misleading.  What is meant by
"passed through to the database"? Why is it described here? Is it
important? Let's remove the paragraph to avoid users having these
questions to begin with.  It's described later on anyway.